rename: pages 아래에 있던 라우팅 컴포넌트를, 데이터 조회만 하도록 변경 및 features 하위에 pages 폴더 생성
